Replace SierraCheckDigits with SierraRecordNumbers
