+++ catalogue_pipeline/transformer/src/main/scala/uk/ac/wellcome/platform/transformer/transformers/SierraTransformableTransformer.scala
@@ -1,5 +1,9 @@
 package uk.ac.wellcome.platform.transformer.transformers
 
+import uk.ac.wellcome.models.transformable.sierra.{
+  SierraRecordNumbers,
+  SierraRecordTypes
+}
 import uk.ac.wellcome.models.transformable.{SierraTransformable, Transformable}
 import uk.ac.wellcome.models.work.internal._
 import uk.ac.wellcome.platform.transformer.source.SierraBibData
@@ -33,7 +37,7 @@ class SierraTransformableTransformer
       val sourceIdentifier = SourceIdentifier(
         identifierType = IdentifierType("sierra-system-number"),
         ontologyType = "Work",
-        value = addCheckDigit(
+        value = SierraRecordNumbers.addCheckDigit(
           sierraTransformable.sourceId,
           recordType = SierraRecordTypes.bibs
         )
